16.16 fixed point values really have more precision than float. Must use double to get bit-perfect result

+++ SwfSharp/Filters/BevelFilter.cs
@@ -11,8 +11,8 @@ namespace SwfSharp.Filters
     {
         public RgbaStruct ShadowColor { get; set; }
         public RgbaStruct HighlightColor { get; set; }
-        public float Angle { get; set; }
-        public float Distance { get; set; }
+        public double Angle { get; set; }
+        public double Distance { get; set; }
         public float Strength { get; set; }
         public bool InnerShadow { get; set; }
         public bool Knockout { get; set; }
